pyproject.toml
[project]
name = "resume-writer"
version = "0.1.0"
license = { file = "LICENSE" }
authors = [
    {name = "Michael Aguilar", email = "bitflip@slowlink.net"},
]
requires-python = "<4.0,>=3.10"
dependencies = [
    "mistune<4.0.0,>=3.0.2",
    "python-docx<2.0.0,>=1.1.2",
    "click<9.0.0,>=8.1.7",
    "tomli<3.0.0,>=2.0.1",
    "dateparser<2.0.0,>=1.2.0",
    "jinja2<4.0.0,>=3.1.4",
    "pytz<2025.0,>=2024.2",
    "nltk<4.0.0,>=3.9.1",
    "mosestokenizer<2.0.0,>=1.2.1",
    "dotenv>=0.9.9",
    "rich>=14.1.0",
    "aider-chat>=0.85.5",
]
description = ""
readme = "README.md"

[tool.ruff]
exclude = ["temporal-samples-python"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "D201", "D202",
    "RET504", # ignore assignment right before return
    "PGH003", # ignore general mypy type: ignore
    "PTH123", # ignore Path.open
    "FA102", # ignore annotation warning, we require v3.10
    "ANN102", # ignore `cls` annotations
    "TD002", "TD003", "FIX002", # ignore TODO details
    "S101", # ignore asserts
    "ANN204", # ignore __init__ return type
    "UP032", "G001", # I like str.format
    "TRY003", # Ignore long exception messages
    "D104", "D100", # Ignore docstring for __init__.py, __main__.py
    "EM101", "EM102", "G004", # Ignore strings in exception and logs, I'm feeling lazy
    "FBT001", # ignore boolean parameters
    "FBT002", # ignore boolean init
    ]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*" = ["ANN", # ignore annotations in tests
                    "D", # ignore docstrings in tests
                    "I", # ignore import order in tests
                    "PLR2004", # ignore magic values in tests
                    "DTZ005", # ignore datetime.now() in tests,
                    "DTZ001", # ignore datetime.utcnow() in tests
                    "B010" # ignore setattr in tests
]

[tool.pytest.ini_options]
pythonpath = [".", "./resume_writer", "./tests"]
addopts = "-rx"
log_cli_level = "DEBUG"

[dependency-groups]
dev = [
    "mypy<2.0.0,>=1.11.0",
    "pytest<9.0.0,>=8.3.1",
    "coverage<8.0.0,>=7.6.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]
